+++ code/extlibs/nvidia-texture-tools/src/nvmath/nvmath.h
@@ -219,7 +219,7 @@ namespace nv
 
     inline float frac(float f)
     {
-        return f - floor(f);
+        return f - floorf(f);
     }
 
     inline float floatRound(float f)
